requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
uvicorn[standard]>=0.27.0
channels>=4.0.0
channels-redis>=4.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
requests>=2.31.0
Pillow>=10.2.0
gunicorn>=21.2.0
whitenoise>=6.6.0
django-storages>=1.14.2
boto3>=1.34.0
django-filter>=23.5
supabase>=2.3.0
pyjwt
supabase
openai>=1.0.0
